nginx-common.nginx.service: workaround for a warning message
Workaround for systemd pidfile handling not to print this warning in the log:
"nginx.service: Failed to read PID from file /run/nginx.pid: Invalid argument"
The warning message shows there is a race condition. The pidfile is not written
yet when systemd is trying to read it.

It is a known bug in ubuntu:
https://bugs.launchpad.net/ubuntu/+source/nginx/+bug/1581864
And this is the suggested workaround.

It was the same fix in balabit-os-6.0:
https://git.balabit/balabit-os-6.0/nginx/commit/f7dd4f41f922531baba7079eddd657dab0e85b25